+++ src/main/kotlin/io/verik/intellij/annotate/VerikAnnotator.kt
@@ -19,10 +19,14 @@ package io.verik.intellij.annotate
 import com.intellij.lang.annotation.AnnotationHolder
 import com.intellij.lang.annotation.Annotator
 import com.intellij.lang.annotation.HighlightSeverity
+import com.intellij.openapi.util.TextRange
 import com.intellij.psi.PsiElement
 import org.jetbrains.kotlin.idea.highlighter.KotlinHighlightingColors
 import org.jetbrains.kotlin.lexer.KtSingleValueToken
 import org.jetbrains.kotlin.psi.KtOperationReferenceExpression
+import org.jetbrains.kotlin.psi.KtTypeElement
+import org.jetbrains.kotlin.psi.psiUtil.endOffset
+import org.jetbrains.kotlin.psi.psiUtil.startOffset
 
 class VerikAnnotator : Annotator {
 
@@ -34,6 +38,28 @@ class VerikAnnotator : Annotator {
                     .newSilentAnnotation(HighlightSeverity.INFORMATION)
                     .textAttributes(KotlinHighlightingColors.EXTENSION_FUNCTION_CALL)
                     .create()
+            }
+        }
+        if (element is KtTypeElement) {
+            val text = element.text
+            if (text.startsWith("`") && text.endsWith("`")) {
+                val int = text.substring(1, text.length - 1).toIntOrNull()
+                if (int != null) {
+                    val startOffset = element.startOffset
+                    val endOffset = element.endOffset
+                    val startTextRange = TextRange(startOffset, startOffset)
+                    val endTextRange = TextRange(startOffset + 1, endOffset)
+                    holder
+                        .newSilentAnnotation(HighlightSeverity.INFORMATION)
+                        .textAttributes(KotlinHighlightingColors.NUMBER)
+                        .range(startTextRange)
+                        .create()
+                    holder
+                        .newSilentAnnotation(HighlightSeverity.INFORMATION)
+                        .textAttributes(KotlinHighlightingColors.NUMBER)
+                        .range(endTextRange)
+                        .create()
+                }
             }
         }
     }